Documenatation and cleanup
